+++ CHANGELOG.md
@@ -11,6 +11,7 @@ master
 * CI: change symfony version from 4.2 to 4.4
 * CI: add php version 7.4 at nightly
 * Upgrade symfony/framework-bundle version to fix security issue
+* Fix deprecations related to block
 
 v0.0.1
 ------