+++ CHANGELOG.md
@@ -12,6 +12,7 @@ master
 * CI: add php version 7.4 at nightly
 * Upgrade symfony/framework-bundle version to fix security issue
 * Adapted ImmutableTabsType and its test class to new Sonata form component (fixed namespace and inheritance)
+* Upgraded ekino/phpstan-banned-code to 0.3 and upgraded its dependencies
 
 v0.0.1
 ------